+++ ext/log-processor/src/test/java/core/diagram/service/ArchDiagramTest.java
@@ -3,8 +3,6 @@ package core.diagram.service;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import java.util.Set;
-
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
@@ -15,7 +13,7 @@ class ArchDiagramTest {
 
     @BeforeEach
     void createArchDiagram() {
-        diagram = new ArchDiagram(Set.of());
+        diagram = new ArchDiagram();
     }
 
     @Test